requirements.txt
streamlit
openai
nltk
pypdf
python-dotenv
requests
langchain
langchain-text-splitters
langchain-openai
sentence-transformers
langchain-chroma
